+++ pysollib/tile/selecttile.py
@@ -130,9 +130,9 @@ class SelectTileDialogWithPreview(MfxDialog):
         sw = self.top.winfo_screenwidth()
         sh = self.top.winfo_screenheight()
 
-        h = sh * .8
-        w = sw * .8
-        w1 = min(275, sw / 2.5)
+        h = int(sh * .8)
+        w = int(sw * .8)
+        w1 = int(min(275, sw / 2.5))
         geometry = ("%dx%d+%d+%d" % (w, h, (sw - w) / 2, (sh - h) / 2))
         self.top.wm_minsize(400, 200)
 
@@ -142,7 +142,7 @@ class SelectTileDialogWithPreview(MfxDialog):
         frame.pack(fill='both', expand=True,
                    padx=kw.padx-padx, pady=kw.pady-pady)
         self.tree = self.Tree_Class(self, frame, key=key, default=kw.default,
-                                    font=font, width=int(w1))
+                                    font=font, width=w1)
         self.tree.frame.pack(side="left", fill='both', expand=False,
                              padx=padx, pady=pady)
         self.preview = MfxScrolledCanvas(frame, hbar=0, vbar=0)
